Correct p-value definition in determine-winner.md
The previous wording suggests that we are providing a definition for the p-value, and that definition was incorrect. I have added a correct definition. Corrected the second sentence to more precisely represent what a p-value can be used for.

Also added a wording change in sentence following the earlier "looser" definition.