COR-4129-clarify AutoAllocate time to confidence
As a followup to COR-4129, clarifying some vague wording about the time it takes for an AutoAllocate activity to start showing confidence. Previous wording could have been interpreted as saying Auto-Allocate would reach 95% confidence (i.e. be complete) in half the time a normal A/B test would run; instead the statement is that "typically, confidence levels are shown after half the time that a normal A/B test would run"